Update index.md
Added
"If you already have installed rust you can look at an offline version of the 
rust documentation by typing “rustup doc” into the console."
on line 29-30 so new rust learners don't make the same mistake I made and try to download the docs by printing to pdf and instead 
know that they just have to type "rustup doc" into the console to have an offline version of it.
I know that almost everyone knows about this but when I learn a new programming language I look at the docs first and somehow missed that. So this minor tweak seemed like a good change to me.